+++ Renderer/RendererBackend/OpenGLES2Renderer/src/Buffer/VertexArrayVao.cpp
@@ -109,13 +109,13 @@ namespace OpenGLES2Renderer
 
 		#ifndef OPENGLES2RENDERER_NO_STATE_CLEANUP
 			// Be polite and restore the previous bound OpenGL ES 2 vertex array
-			glBindVertexArrayOES(openGLES2VertexArrayBackup);
+			glBindVertexArrayOES(static_cast<GLuint>(openGLES2VertexArrayBackup));
 
 			// Be polite and restore the previous bound OpenGL ES 2 element array buffer
-			glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, openGLES2ElementArrayBufferBackup);
+			glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, static_cast<GLuint>(openGLES2ElementArrayBufferBackup));
 
 			// Be polite and restore the previous bound OpenGL ES 2 array buffer
-			glBindBuffer(GL_ARRAY_BUFFER, openGLES2ArrayBufferBackup);
+			glBindBuffer(GL_ARRAY_BUFFER, static_cast<GLuint>(openGLES2ArrayBufferBackup));
 		#endif
 	}
