+++ Renderer/RendererBackend/OpenGLES2Renderer/src/OpenGLES2Renderer.cpp
@@ -540,7 +540,7 @@ namespace OpenGLES2Renderer
 
 							#ifndef OPENGLES2RENDERER_NO_STATE_CLEANUP
 								// Be polite and restore the previous active OpenGL ES 2 texture
-								glActiveTexture(openGLES2ActiveTextureBackup);
+								glActiveTexture(static_cast<GLuint>(openGLES2ActiveTextureBackup));
 							#endif
 							break;
 						}
@@ -1325,7 +1325,7 @@ namespace OpenGLES2Renderer
 
 				#ifndef OPENGLES2RENDERER_NO_STATE_CLEANUP
 					// Be polite and restore the previous bound OpenGL ES 2 array element buffer
-					glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, openGLES2ArrayElementBufferBackup);
+					glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, static_cast<GLuint>(openGLES2ArrayElementBufferBackup));
 				#endif
 
 				// Done
@@ -1359,7 +1359,7 @@ namespace OpenGLES2Renderer
 
 				#ifndef OPENGLES2RENDERER_NO_STATE_CLEANUP
 					// Be polite and restore the previous bound OpenGL ES 2 array buffer
-					glBindBuffer(GL_ARRAY_BUFFER, openGLES2ArrayBufferBackup);
+					glBindBuffer(GL_ARRAY_BUFFER, static_cast<GLuint>(openGLES2ArrayBufferBackup));
 				#endif
 
 				// Done
@@ -1491,7 +1491,7 @@ namespace OpenGLES2Renderer
 
 				#ifndef OPENGLES2RENDERER_NO_STATE_CLEANUP
 					// Be polite and restore the previous bound OpenGL ES 2 array element buffer
-					glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, openGLES2ArrayElementBufferBackup);
+					glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, static_cast<GLuint>(openGLES2ArrayElementBufferBackup));
 				#endif
 				break;
 			}
@@ -1519,7 +1519,7 @@ namespace OpenGLES2Renderer
 
 				#ifndef OPENGLES2RENDERER_NO_STATE_CLEANUP
 					// Be polite and restore the previous bound OpenGL ES 2 array buffer
-					glBindBuffer(GL_ARRAY_BUFFER, openGLES2ArrayBufferBackup);
+					glBindBuffer(GL_ARRAY_BUFFER, static_cast<GLuint>(openGLES2ArrayBufferBackup));
 				#endif
 				break;
 			}
